In all MPI_Bcast type of nbSteps is sent as MPI_UNSIGNED_LONG. Caused failures when mixed MPI_INT and MPI_UNSIGNED_LONG.

+++ src/main.cpp
@@ -60,7 +60,7 @@ static void master(std::string filename)
     while (!engine.done(cumData))
     {
         /* Send the slave a new work unit */
-        MPI_Bcast(&curData->m_nbSteps, 1, MPI_INT, 0, MPI_COMM_WORLD);
+        MPI_Bcast(&curData->m_nbSteps, 1, MPI_UNSIGNED_LONG, 0, MPI_COMM_WORLD);
 
         /*master also works*/
         calculator->run(curData);
